Panel 분리
