+++ Client/src/client/ui/userview/UserView.java
@@ -1,4 +1,4 @@
-package client.ui;
+package client.ui.userview;
 
 import client.data.dao.ProductModel;
 
@@ -12,7 +12,7 @@ public interface UserView {
     public JPanel startPnl = new JPanel(), itemListPnl = new JPanel(), selectedListPnl = new JPanel();
     public JButton btnAdminClient = new JButton(), btnPay = new JButton("구매"), btnMinus = new JButton("-"), btnPlus = new JButton("+"), btnX = new JButton("x");
     public JLabel lblInsertMoney = new JLabel("투입 금액 "), lblTotalMoney = new JLabel("총 금액 "), lblChangeMoney = new JLabel("거스름돈 "), lblItemQuantity = new JLabel("3");
-    public List<UserViewImpl.ItemInfoPnl> itemLists = new LinkedList<>();
+    public List<ItemInfoPnl> itemLists = new LinkedList<>();
 
     // TODO: 2019-12-31 파라미터
     void updateItemLists(Vector<ProductModel> lists);
